List command unified & localized.
